+++ src/Models/Channel/Scopes/HideFuture.php
@@ -1,0 +1,20 @@
+<?php
+
+namespace Expressionengine\Coilpack\Models\Channel\Scopes;
+
+use Illuminate\Database\Eloquent\Builder;
+use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\Scope;
+
+class HideFuture implements Scope
+{
+    /**
+     * Apply the scope to a given Eloquent query builder.
+     *
+     * @return void
+     */
+    public function apply(Builder $builder, Model $model)
+    {
+        $builder->where('entry_date', '<=', now()->timestamp);
+    }
+}